remarks taken into account, payment component moved to the components folder
