+++ client/src/pages/MastersPage/MastersPage.jsx
@@ -73,9 +73,6 @@ const MastersPage = () => {
         <Route path="/masters/order">
           <OrderForm onSubmit={handleSubmitOrder} />
         </Route>
-        {/* <Route path="/masters/payment">
-          <PaymentBlock/>
-        </Route> */}
       </Switch>
     </div>
   )
